Errors corregits

+++ src/Gestio/Empleats/Eliminar_empleats.java
@@ -54,19 +54,28 @@ public class Eliminar_empleats extends javax.swing.JFrame{
                     JOptionPane.showMessageDialog(frame_eliminarEmpleats, "No hi han empleats a eliminar");
                 }else{
                     try{
-                        for (int i=0; i<Persona.getContador(); i++){
-                            if (indices2.toString().equalsIgnoreCase(Persona.arrayPersones[i].toString())){
-                                IO.imprimirTI("Eliminar index "+i);
-                                for ( int k = i ; k < Persona.getContador() - 1 ; k++)
-                                {
-                                    Persona.arrayPersones[ k ] = Persona.arrayPersones[ k + 1 ] ;
+                        int dialogButton = JOptionPane.YES_NO_OPTION;
+                        int dialogResult = JOptionPane.showConfirmDialog(null, "Estas segur de que vols eliminar el empleat?", "CONFIRMACIO", dialogButton);
+                        if(dialogResult == 0) {
+                            IO.imprimirTI("Opcio 'SI'");
+                            for (int i=0; i<Persona.getContador(); i++){
+                                if (indices2.toString().equalsIgnoreCase(Persona.arrayPersones[i].toString())){
+                                    IO.imprimirTI("Eliminar index "+i);
+                                    for ( int k = i ; k < Persona.getContador() - 1 ; k++)
+                                    {
+                                        Persona.arrayPersones[ k ] = Persona.arrayPersones[ k + 1 ] ;
+                                    }
+                                    Persona.eliminarPersona();
+                                    IO.imprimirTI("Empleat eliminat.");
+                                    JOptionPane.showMessageDialog(frame_eliminarEmpleats, "Empleat eliminat correctament");
+                                    Auxiiliar.llistar_empleats(cercaEmpleat, jList2);
                                 }
-                                Persona.eliminarPersona();
-                                IO.imprimirTI("Empleat eliminat.");
-                                JOptionPane.showMessageDialog(frame_eliminarEmpleats, "Empleat eliminat correctament");
-                                Auxiiliar.llistar_empleats(cercaEmpleat, jList2);
                             }
+                        } else {
+                            IO.imprimirTI("Opcio 'NO'");
                         }
+
+
 
                     }catch (Exception error){
                         IO.imprimirTI("Error: "+error);
